Move QbftContext out of the pki package as it is also used in non-pki mode
